chore: remove versioning code and related files
Removed code and files related to API versioning as it is not currently adopted. The versioning-related components, such as the custom mapper (`TeacherMapper`), methods (`createV2`), DTOs, and endpoint specific to version 2 of the API, have been deleted. These changes were made for study purposes, and at the moment, API versioning is not being adopted. This cleanup simplifies the codebase and removes unnecessary components.